requirements.txt
alembic==1.4.2
apache-airflow==1.10.10
apispec==1.3.3
argcomplete==1.11.1
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
boto3==1.12.36
botocore==1.15.36
cached-property==1.5.1
cassandra-driver==3.22.0
cattrs==0.9.0
certifi==2019.11.28
chardet==3.0.4
click==6.7
colorama==0.4.3
colorlog==4.0.2
configparser==3.5.3
croniter==0.3.31
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.1.1
docutils==0.16
Flask==1.1.1
Flask-Admin==1.5.4
Flask-AppBuilder==2.3.0
Flask-Babel==1.0.0
Flask-Caching==1.3.3
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.1
flask-swagger==0.2.13
Flask-WTF==0.14.3
funcsigs==1.0.2
future==0.18.2
geomet==0.1.2
graphviz==0.13.2
gunicorn==19.10.0
idna==2.9
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipython-sql==0.3.9
iso8601==0.1.12
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.10.3
jmespath==0.9.5
json-merge-patch==0.2
jsonschema==3.2.0
jupyter-client==6.0.0
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
Mako==1.1.2
Markdown==2.6.11
MarkupSafe==1.1.1
marshmallow==2.19.5
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.18.0
numpy==1.18.2
pandas==0.25.3
parso==0.6.2
pendulum==1.4.4
pexpect==4.8.0
pickleshare==0.7.5
prettytable==0.7.2
prison==0.1.3
prompt-toolkit==3.0.4
psutil==5.7.0
psycopg2-binary==2.8.4
ptyprocess==0.6.0
Pygments==2.6.1
PyJWT==1.7.1
pyrsistent==0.15.7
python-daemon==2.1.2
python-dateutil==2.8.1
python-editor==1.0.4
python3-openid==3.1.0
pytz==2019.3
pytzdata==2019.3
PyYAML==5.3.1
pyzmq==19.0.0
requests==2.23.0
s3transfer==0.3.3
setproctitle==1.1.10
six==1.14.0
SQLAlchemy==1.3.15
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.36.3
sqlparse==0.3.1
tabulate==0.8.6
tenacity==4.12.0
termcolor==1.1.0
text-unidecode==1.2
thrift==0.13.0
tornado==6.0.4
traitlets==4.3.3
typing==3.7.4.1
typing-extensions==3.7.4.1
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.25.8
wcwidth==0.1.8
Werkzeug==0.16.1
WTForms==2.2.1
zipp==1.2.0
zope.deprecation==4.4.0
